requirements.txt
#pydantic>=1.8.2
#python-jose==3.3.0
#email-validator==2.1.0.post1
#python-multipart==0.0.6
#uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
fastapi==0.115.11
sqlalchemy==2.0.23
Authlib==1.2.1
httpx==0.25.1
python-jose==3.4.0
passlib==1.7.4
argon2-cffi==23.1.0
psycopg2-binary==2.9.10
email-validator==2.2.0
python-multipart==0.0.20
uvicorn==0.34.0

#
#
#alembic==1.12.1
#psycopg2-binary==2.9.9
#bcrypt==4.0.1
#
#requests==2.31.0
#python-dotenv==1.0.0
#
#starlette==0.27.0
#UUID==1.30
#rate-limiter==0.0.2
#fastapi>=0.68.0
#uvicorn>=0.15.0
#sqlalchemy>=1.4.23
#python-jose[cryptography]>=3.3.0
#passlib[bcrypt]>=1.7.4
#mangum>=0.15.0
#azure-functions>=1.12.0